Add Repos for all models and did some correction in primary keys of m…

+++ Backend/src/main/java/com/auto/exam/Model/CourseRegister.java
@@ -9,7 +9,7 @@ public class CourseRegister {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "course_registration_id")
-    private Integer courseRegistrationID;
+    private Long courseRegistrationID;
 
     @ManyToOne
     @JoinColumn(name = "registration_id", nullable = false)
